+++ Tests/Realm.Tests/Sync/MongoClientTests.cs
@@ -797,6 +797,9 @@ namespace Realms.Tests.Sync
             {
                 var collection = await GetCollection();
 
+                var nonexistent = await collection.FindOneAsync();
+                Assert.That(nonexistent, Is.Null);
+
                 var inserted = await InsertSomeData(collection, 3);
 
                 var result = await collection.FindOneAsync();
@@ -2105,7 +2108,8 @@ namespace Realms.Tests.Sync
                 var projection = BsonDocument.Parse("{ LongValue: 1 }");
 
                 var result = await collection.FindOneAndDeleteAsync(filter, sort, projection);
-                Assert.That(result.Id, Is.EqualTo(inserted[1].Id));
+                Assert.That(result, Is.Not.Null);
+                Assert.That(result!.Id, Is.EqualTo(inserted[1].Id));
                 Assert.That(result.StringValue, Is.Null);
                 Assert.That(result.LongValue, Is.EqualTo(inserted[1].LongValue));
 
@@ -2123,7 +2127,8 @@ namespace Realms.Tests.Sync
                 var inserted = await InsertRemappedData(collection);
 
                 var result = await collection.FindOneAsync();
-                Assert.That(result.Id, Is.EqualTo(inserted[0].Id));
+                Assert.That(result, Is.Not.Null);
+                Assert.That(result!.Id, Is.EqualTo(inserted[0].Id));
                 Assert.That(result.StringValue, Is.EqualTo(inserted[0].StringValue));
                 Assert.That(result.MappedLink!.Id, Is.EqualTo(inserted[1].Id));
                 Assert.That(result.MappedList[0].Id, Is.EqualTo(inserted[2].Id));